Moving reusable parts of chakra-react-select to separate files. Also fixes stlying issues with MenuList scrollbar and height overflowing page
